+++ packages/skills/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @agent-native/skills
 
+## 0.2.457
+
+### Patch Changes
+
+- Updated dependencies [9258da4]
+  - @agent-native/core@0.133.3
+
+## 0.2.456
+
+### Patch Changes
+
+- Updated dependencies [3fac05d]
+  - @agent-native/core@0.133.2
+
+## 0.2.455
+
+### Patch Changes
+
+- Updated dependencies [1c08605]
+  - @agent-native/core@0.133.1
+
 ## 0.2.454
 
 ### Patch Changes